update fwknop to 2.5.1-1 includes head changes
A new minor version of fwknop was released today. This contains the
fix for fwknop not building on OS X Mavericks (10.9).

In addition to updating the URL and updating the SHA1SUM, some modifications
were needed so that the GitHub release could be built. This includes setting
dependencies for the things needed by autogen.sh, as well as running autogen.sh
when it is needed.

Closes #24419.

Signed-off-by: BrewTestBot <brew-test-bot@googlegroups.com>