ruby-build: get rid of libyaml blocker on install
libyaml is recommended to have for installing *Rubies*, not ruby-build
itself. Obviously, Homebrew is unaware of this distinction and fails
when I tried to `brew install ruby-build`:

  Error: You must `brew link libyaml' before ruby-build can be installed

I have libyaml installed, but it's not linked. It doesn't need to be
linked for ruby-build to automatically use it while compiling Rubies.