colormake 0.9 not as head-only
tarball for colormake-0.9 is still available from the author
(even if the future of colormake is now on Github),
then this formula gives this tool for all users without switching to head-only.

Besides, it adds the manpages and all variants of the bin commands not supplied by the current head-only formula.