Update to wv-1.2.9 to fix crashes with plone.
The wv-1.2.6 version is known to crash when used with Plone to make
Word documents and others searchable. This issues was identified in
debian and they fixed by upgrading to 1.2.9-2 (the -2 is the debian
packaging version.

The initial built failed due to some incompatibility in the given
install-sh script used to install the files. This is a known issue
as evidenced by this message in the abiword mailing list:

http://www.abisource.com/mailinglists/abiword-dev/2011/Jun/0108.html

A workaround has been implemented by creating the folders before
calling make install. This fixes the installation issue till it is
fixed upstream.