Add support for compiling QuantLib with --c++11
Although QuantLib does not have any special features/support for C++11,
it is still useful for developers/administrators who need to link
against libc++ and to combine QuantLib with other C++11 applications and
libraries.

Also be sure to depend on boost with c++11 to ensure that QuantLib and
boost share the same standard library.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>