python: pip executable be at HOMEBREW_PREFIX/bin
Removed the `--install-script=#{bin}"` arg to the python setup.py call
and also move that call _after_ we wrote the distutils.cfg. The reason
for this is that `pip` should end up in `HOMEBREW_PREFIX/bin/`. This is
no longer the case since the install of pip was moved to `postinstall`
because there is no more symlinking from cellar to prefix.